Move service/parserprovider package to config/configmapprovider
