+++ config/configmapprovider/default.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package parserprovider // import "go.opentelemetry.io/collector/service/parserprovider"
+package configmapprovider // import "go.opentelemetry.io/collector/config/configmapprovider"
 
 import "go.opentelemetry.io/collector/config"
