+++ service/command.go
@@ -17,7 +17,7 @@ package service // import "go.opentelemetry.io/collector/service"
 import (
 	"github.com/spf13/cobra"
 
-	"go.opentelemetry.io/collector/service/parserprovider"
+	"go.opentelemetry.io/collector/config/configmapprovider"
 )
 
 // NewCommand constructs a new cobra.Command using the given Collector.
@@ -29,7 +29,7 @@ func NewCommand(set CollectorSettings) *cobra.Command {
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if set.ConfigMapProvider == nil {
-				set.ConfigMapProvider = parserprovider.NewDefaultMapProvider(getConfigFlag(), getSetFlag())
+				set.ConfigMapProvider = configmapprovider.NewDefaultMapProvider(getConfigFlag(), getSetFlag())
 			}
 			col, err := New(set)
 			if err != nil {
